Convert PrivacySettingsDialog to TypeScript
Unfortunately type checks fail because TS finds errors in `node_modules/react-i18next/src/index.d.ts`
